backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
python-decouple==3.8
pymongo==4.6.0
mongoengine==0.27.0
django-mongoengine==0.5.6
bcrypt==4.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
